DynamicDiscreteSamplers

Documentation for DynamicDiscreteSamplers.

DynamicDiscreteSamplers.FixedSizeWeights

DynamicDiscreteSamplers.ResizableWeights

DynamicDiscreteSamplers.SemiResizableWeights

DynamicDiscreteSamplers.Weights

DynamicDiscreteSamplers._resize!

DynamicDiscreteSamplers.binomial_sample

DynamicDiscreteSamplers.binomial_sample_fair_coin

DynamicDiscreteSamplers.initialize_empty

DynamicDiscreteSamplers.multinomial_sample

DynamicDiscreteSamplers.FixedSizeWeights — Type

FixedSizeWeights <: Weights

An object that confomrs the the Weights interface and cannot be resized.

DynamicDiscreteSamplers.ResizableWeights — Type

ResizableWeights <: Weights

An object that confomrs the the Weights interface and can be resized.

DynamicDiscreteSamplers.SemiResizableWeights — Type

SemiResizableWeights <: Weights

An object that confomrs the the Weights interface and can be resized, but only to sizes at most as large as it's original size.

DynamicDiscreteSamplers.Weights — Type

Weights <: AbstractVector{Float64}

An abstract vector capable of storing normal, non-negative floating point numbers on which rand samples an index according to values rather than sampling a value uniformly.

DynamicDiscreteSamplers._resize! — Method

Reallocate w with the size len, compacting w into that new memory. Any elements if w past len must be set to zero already (that's a general invariant for Weights, though, not just this function).

DynamicDiscreteSamplers.binomial_sample — Method

binomial_sample(rng, trials, px, py)

Flip a coin with probability of px//py of coming up heads trials times and return the number of heads.

Has O(trials) expected runtime with a very low constant factor.

Implementation based on Farach-Colton, M. and Tsai, M.T., 2015. Exact sublinear binomial sampling.

DynamicDiscreteSamplers.binomial_sample_fair_coin — Method

binomial_sample_fair_coin(rng, trials)

Flips trials fair coins and reports the number of heads.

Flips up to 64 coins at a time.

DynamicDiscreteSamplers.initialize_empty — Method

initialize_empty(len::Int)::Memory{UInt64}

Initialize a Memory that, when underlaying a Weights object, represents len zeros.

DynamicDiscreteSamplers.multinomial_sample — Method

multinomial_sample(rng, trials, weights)

Draw trials elements from the probability distribution specified by weights (need not sum to 1) and return the number of times each element was drawn.

Runs in O(trials * weights), but can be as fast as O(trials) if the weights are skewed toward big weights at the beginning.